Move executable bootstrap sources into src/source/App/
Relocate the executable shell's files (stdafx.h, StdAfx.cpp, Winmain.cpp,
Winmain.h, and the Platform/Windows resource and platform files) into a
dedicated src/source/App/ directory so the folder maps to the sources
compiled into the executable target after the library/executable split.

Keep source/App on the engine library's PUBLIC include path so the bare
include "stdafx.h" in roughly 350 translation units resolves unchanged,
and rewrite the explicit Platform/Windows header includes to their new
App/ paths. Update the executable target to compile source/App/Winmain.cpp
and source/App/Platform/Windows/resource.rc.
