+++ src/source/App/Platform/Windows/icon2.ico
