+++ src/source/App/Platform/Windows/resource.h
