+++ src/source/App/Platform/Windows/resource.res
