+++ src/source/Engine/Object/ZzzInventory.cpp
@@ -21,7 +21,7 @@
 
 #include "Core/Utilities/Debouncer.h"
 #include "GameLogic/Quests/CSQuest.h"
-#include "Platform/Windows/Local.h"
+#include "App/Platform/Windows/Local.h"
 #include "GameLogic/Items/PersonalShopTitleImp.h"
 #include "Engine/AI/GOBoid.h"
 #include "GameLogic/Items/CSItemOption.h"
