+++ src/source/Scenes/SceneManager.cpp
@@ -41,7 +41,7 @@ FrameTimingState g_frameTiming;
 #include "UI/NewUI/HUD/Notices.h"
 #include "I18N/All.h"
 #include "Engine/AI/ZzzAI.h"
-#include "Platform/Windows/Winmain.h"
+#include "App/Winmain.h"
 #include "Camera/CameraManager.h"
 #include "Camera/CameraMode.h"
