+++ src/source/UI/NewUI/HUD/NewUIMasterLevel.cpp
@@ -1,5 +1,5 @@
 ﻿#include "stdafx.h"
-#include "Platform/Windows/Winmain.h"
+#include "App/Winmain.h"
 #include "Render/Textures/ZzzTexture.h"
 #include "GameLogic/Items/CSItemOption.h"
 #include "UI/Legacy/UIControls.h"
